ユーザ登録

+++ 2311teamDevelopment/src/main/java/com/example/demo/repository/UserRegistrationRepository.java
@@ -6,4 +6,4 @@ import org.springframework.stereotype.Repository;
 import com.example.demo.entity.UserRegistrationEntity;
 
 @Repository
-public interface UserRegistrationRepository extends JpaRepository<UserRegistrationEntity, Long> {}
+public interface UserRegistrationRepository extends JpaRepository<UserRegistrationEntity, Integer> {}
